requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
anthropic>=0.7.0
httpx>=0.24.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10

